requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
gunicorn
pymongo==4.9.1
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dateutil
requests
